refactor: Refactor structure of Frontend
Repair index.tsx files

Refs: CU-86975w8k1
