+++ frontend/src/Features/Login/components/Footer/index.tsx
@@ -1,1 +1,1 @@
-export * from './Footer.tsx';
+export * from './Footer';
