+++ frontend/src/Features/Messenger/components/Chat/index.tsx
@@ -1,1 +1,1 @@
-export * from './Chat.tsx';
+export * from './Chat';
