+++ frontend/src/Features/Messenger/components/MessageOptions/index.tsx
@@ -1,1 +1,1 @@
-export * from './MessageOptions.tsx';
+export * from './MessageOptions';
