+++ frontend/src/Features/Messenger/components/MessageSender/index.tsx
@@ -1,1 +1,1 @@
-export * from './MessageSender.tsx';
+export * from './MessageSender';
