+++ frontend/src/Features/Messenger/components/UnselectedChat/index.tsx
@@ -1,1 +1,1 @@
-export * from './UnselectedChat.tsx';
+export * from './UnselectedChat';
