+++ frontend/src/Features/Profile/components/ProfileFriends/index.tsx
@@ -1,1 +1,1 @@
-export * from './ProfileFriends.tsx'
+export * from './ProfileFriends'
