+++ frontend/src/Features/Profile/components/ProfileMultimedia/index.tsx
@@ -1,1 +1,1 @@
-export * from './ProfileMultimedia.tsx'
+export * from './ProfileMultimedia'
