+++ frontend/src/Features/Root/components/Buttons/LogOut/index.tsx
@@ -1,1 +1,1 @@
-export * from './LogOut.tsx';
+export * from './LogOut';
