+++ frontend/src/Features/Root/components/Buttons/Menu/index.tsx
@@ -1,1 +1,1 @@
-export * from './MenuButton.tsx';
+export * from './MenuButton';
