+++ frontend/src/Features/Root/components/Layout/index.tsx
@@ -1,1 +1,1 @@
-export * from "./Layout.tsx"
+export * from "./Layout"
