+++ frontend/src/Features/Root/components/Navbar/index.tsx
@@ -1,1 +1,1 @@
-export * from './Navbar.tsx';
+export * from './Navbar';
