+++ frontend/src/Features/Root/components/ProtectedRoute/index.tsx
@@ -1,1 +1,1 @@
-export * from './ProtectedRoute.tsx';
+export * from './ProtectedRoute';
