+++ frontend/src/Features/Settings/components/Escape/index.tsx
@@ -1,1 +1,1 @@
-export * from './Escape.tsx';
+export * from './Escape';
