+++ frontend/src/Features/Settings/components/Menu/About/index.tsx
@@ -1,1 +1,1 @@
-export * from './About.tsx';
+export * from './About';
