+++ frontend/src/Features/Settings/components/Menu/Accessibility/index.tsx
@@ -1,1 +1,1 @@
-export * from './Accessibility.tsx';
+export * from './Accessibility';
