+++ frontend/src/Features/Settings/components/Menu/Apperance/index.tsx
@@ -1,1 +1,1 @@
-export * from './Apperance.tsx';
+export * from './Apperance';
