+++ frontend/src/Features/Settings/components/Menu/Data/index.tsx
@@ -1,1 +1,1 @@
-export * from './Data.tsx';
+export * from './Data';
