+++ frontend/src/Features/Settings/components/Menu/General/index.tsx
@@ -1,1 +1,1 @@
-export * from './General.tsx';
+export * from './General';
