+++ frontend/src/Features/Settings/components/Menu/Help/index.tsx
@@ -1,1 +1,1 @@
-export * from './Help.tsx';
+export * from './Help';
