+++ frontend/src/Features/Settings/components/Menu/Language/index.tsx
@@ -1,1 +1,1 @@
-export * from './Language.tsx';
+export * from './Language';
