+++ frontend/src/Features/Settings/components/Menu/MyAccount/index.tsx
@@ -1,1 +1,1 @@
-export * from './MyAccount.tsx';
+export * from './MyAccount';
