+++ frontend/src/Features/Settings/components/Menu/Notifications/index.tsx
@@ -1,1 +1,1 @@
-export * from './Notifications.tsx';
+export * from './Notifications';
