+++ frontend/src/Features/Settings/components/Menu/Privacy/index.tsx
@@ -1,1 +1,1 @@
-export * from './Privacy.tsx';
+export * from './Privacy';
