+++ frontend/src/Features/Settings/components/Menu/Security/index.tsx
@@ -1,1 +1,1 @@
-export * from './Security.tsx';
+export * from './Security';
