+++ frontend/src/Features/shared/components/CenterContainer/index.tsx
@@ -1,1 +1,1 @@
-export * from './CenterContainer.tsx';
+export * from './CenterContainer';
