+++ frontend/src/Features/shared/components/ContainerVhVw/index.tsx
@@ -1,1 +1,1 @@
-export * from "./ContainerVhVw.tsx"
+export * from "./ContainerVhVw"
