+++ frontend/src/Features/shared/components/CookiesPopup/index.tsx
@@ -1,1 +1,1 @@
-export * from './CookiesPopup.tsx';
+export * from './CookiesPopup';
