+++ frontend/src/Features/shared/components/RichEditor/index.tsx
@@ -1,1 +1,1 @@
-export * from './RichEditor.tsx';
+export * from './RichEditor';
