+++ frontend/src/Features/shared/consts/index.tsx
@@ -1,0 +1,1 @@
+export {Languages} from './Languages';
