+++ frontend/src/__tests__/Footer.test.tsx
@@ -1,7 +1,7 @@
 import {fireEvent} from '@testing-library/react';
 import '@testing-library/jest-dom';
 import {Footer} from '../Features/Login/components/Footer';
-import {render} from "./test-utils/render";
+import {render} from "../test-utils/render";
 
 const mockChangeLanguage = jest.fn();
